refactor: add cluster-logger.ts
