+++ lib/cluster/cluster-logger.ts
@@ -1,0 +1,4 @@
+import { Logger } from '@nestjs/common';
+import { CLUSTER_MODULE_ID } from './cluster.constants';
+
+export const logger = new Logger(CLUSTER_MODULE_ID);
