+++ lib/cluster/common/index.spec.ts
@@ -1,7 +1,7 @@
 import * as allExports from '.';
 
-test('should have 7 exports', () => {
-    expect(Object.keys(allExports)).toHaveLength(7);
+test('should have 6 exports', () => {
+    expect(Object.keys(allExports)).toHaveLength(6);
 });
 
 test('each of exports should be defined', () => {
